update to new API and remove obsolete code
