+++ CurlEx.curl
@@ -2,6 +2,5 @@
 
 curl -v -X POST "https://api.msrc.microsoft.com/engage/cars"
 -H "Content-Type: application/json"
--H "api-key: {subscription key}"
 
 --data-ascii "{body}"
