+++ phpEx.php
@@ -2,13 +2,12 @@
 // This sample uses the Apache HTTP client from HTTP Components (http://hc.apache.org/httpcomponents-client-ga/)
 require_once 'HTTP/Request2.php';
 
-$request = new Http_Request2('https://api.msrc.microsoft.com/engage/cars');
+$request = new Http_Request2('https://api.msrc.microsoft.com/report/v2.0/abuse');
 $url = $request->getUrl();
 
 $headers = array(
     // Request headers
-    'Content-Type' => 'application/json',
-    'api-key' => '{subscription key}',
+    'Content-Type' => 'application/json'
 );
 
 $request->setHeader($headers);
@@ -21,7 +20,7 @@ $url->setQueryVariables($parameters);
 
 $request->setMethod(HTTP_Request2::METHOD_POST);
 
-// Request body
+// Data model documentation at https://msrc.microsoft.com/report/developer
 $request->setBody("{body}");
 
 try
