--- src/msrcCars/Powershell/Powershell-Body.txt
@@ -1,1 +1,0 @@
-{ "reporterInfo": { "reporterEmail": "yourName@yourOrg.com","reporterName": "Orion Cox", "reporterPhone": "555-555-5555", "reporterOrg": "yourOrg", "discloseEmail": "TRUE", "reporterNotes": "This is example code from GitHub"}, "reports": [ { "report": { "reportNotes": "This is example code from GitHub", "TLP": "WHITE", "disclosureNotes": "This is example code from GitHub", "threats": [ { "threat": { "threatType": "Activity", "threatSubType": "INTRUSION_ATTEMPT", "sampleType": "String", "sourceIp": "1.1.1.1", "destinationIp": "2.2.2.2", "sourcePort": "22", "destinationPort": "1433", "sourceUrl": "attackingUrl.com", "destinationUrl": "attackedUrl.com", "protocol": "2020","byteCount": "100", "packetCount": "95", "date": "2000-01-25", "time": "10:00:00", "timeZone": "-00:00", "sample": "Sample goes here" } } ] } } ] }
