--- src/msrcCars/Powershell/Powershell-Example-Code.Ps1
@@ -1,4 +1,0 @@
-$key = Read-Host 'What is your API Key?'
-$body = Read-Host 'Insert JSON Body here'
-Invoke-WebRequest -UseBasicParsing https://api.msrc.microsoft.com/engage/cars -ContentType "application/json" -Header @{ "api-key" = $key } -Method POST -Body $body
-Read-Host -Prompt "Press Enter to Exit..."
