--- src/msrcCars/Readme.md
@@ -1,38 +1,0 @@
-<h2>Overview</h2>
-
-These code samples show how to use the API for submitting Cloud Abuse Reporting Schema (CARS) reports to to Microsoft's Computer Emergency Response Team (CERT). This endpoint is for reporting suspected cyberattacks or abuse originating from Microsoft Online Services, such as Microsoft Azure, Bing, Outlook, One Drive, and Office 365. This includes malicious network activity originating from a Microsoft IP addresses. It also includes distribution of malicious content or other illicit or illegal material through a Microsoft Online Service.
-
-Note that to use this API (and any of these samples), you will need to log into the [MSRC Portal](https://portal.msrc.microsoft.com/en-us/developer) and obtain an API key on the Developer tab.
-
-<h2>Using the Code Samples</h2>
-
-<h3>C#</h3>
-To use the C# example code, you will need to copy the code into your project or solution because the C# example code file is not a runnable solution. Instructions for creating a new solution are included in the [tutorial](https://github.com/Microsoft/MSRC-Microsoft-Engage-API/blob/barborico/publication-review/src/msrcCars/CsharpExamples/C%23Tutorial.docx).
-
-<h3>Powershell</h3>
-To use the PowerShell module, download the Powershell-Example-Code.Ps1 file and run it using a PowerShell console, command line, or by right-clicking on the file and selecting "Run With Powershell".  Then input your API key and the personalized JSON body that you filled out from the PowershellBody.txt file, and it will attempt to post to the MSRC Engage API.
-
-If you are running this script via the command line, you will need to call it like so:
-
-<b>C:\Users\username> powershell -noexit "yourFilePath"</b>
-
-<h3>Python</h3>
-To use the Python code example, you will need to copy and paste the code into your project because the Python example code file is not runnable as-is.
-
-<h3>Ruby</h3>
-To use the Ruby code example, you will need to copy and paste the code into your project because the Ruby example code file is not runnable as-is.
-
-<h3>Curl</h3>
-To use the Curl code example, you will need to copy and paste the code into your project because the Curl example code file is not runnable as-is.
-
-<h3>Java</h3>
-To use the Java code example, you will need to copy and paste the code into your project because the Java example code file is not runnable as-is.
-
-<h3>JavaScript</h3>
-To use the JavaScript code example, you can copy the code into a text editor that runs JS (e.g., VS Code) or you can download the file and run it using that text editor.
-
-<h3>Objective C</h3>
-To use the Objective C code example, you will need to copy and paste the code into your project because the ObjC example code file is not runnable as-is.
-
-<h3>PHP</h3>
-To use the PHP code sample, you can either copy and paste the code into an existing PHP project, or you can download and run the file as a PHP file with your IDE.
